Todo_Logging/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow==9.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
